+++ packages/expo-av/src/AV.ts
@@ -50,6 +50,7 @@ export const _DEFAULT_INITIAL_PLAYBACK_STATUS: AVPlaybackStatusToSet = {
   rate: 1.0,
   shouldCorrectPitch: false,
   volume: 1.0,
+  audioPan: 0,
   isMuted: false,
   isLooping: false,
 };
@@ -133,6 +134,9 @@ export function assertStatusValuesInBounds(status: AVPlaybackStatusToSet): void 
   }
   if (typeof status.volume === 'number' && (status.volume < 0 || status.volume > 1)) {
     throw new RangeError('Volume value must be between 0.0 and 1.0');
+  }
+  if (typeof status.audioPan === 'number' && (status.audioPan < -1 || status.audioPan > 1)) {
+    throw new RangeError('Pan value must be between -1.0 and 1.0');
   }
 }
 
@@ -313,10 +317,12 @@ export interface Playback extends AV {
   ): Promise<AVPlaybackStatus>;
 
   /**
-   * This is equivalent to `playbackObject.setStatusAsync({ volume })`.
+   * This is equivalent to `playbackObject.setStatusAsync({ volume, audioPan })`.
+   * Note: `audioPan` is currently only supported on Android using `androidImplementation: 'MediaPlayer'`
    * @param volume A number between `0.0` (silence) and `1.0` (maximum volume).
+   * @param audioPan A number between `-1.0` (full left) and `1.0` (full right).
    */
-  setVolumeAsync(volume: number): Promise<AVPlaybackStatus>;
+  setVolumeAsync(volume: number, audioPan?: number): Promise<AVPlaybackStatus>;
 
   /**
    * This is equivalent to `playbackObject.setStatusAsync({ isMuted })`.
@@ -391,8 +397,8 @@ export const PlaybackMixin = {
     });
   },
 
-  async setVolumeAsync(volume: number): Promise<AVPlaybackStatus> {
-    return (this as any as Playback).setStatusAsync({ volume });
+  async setVolumeAsync(volume: number, audioPan?: number): Promise<AVPlaybackStatus> {
+    return (this as any as Playback).setStatusAsync({ volume, audioPan });
   },
 
   async setIsMutedAsync(isMuted: boolean): Promise<AVPlaybackStatus> {