+++ packages/expo-av/src/Audio/Sound.ts
@@ -435,7 +435,7 @@ export class Sound implements Playback {
     shouldCorrectPitch: boolean,
     pitchCorrectionQuality?: PitchCorrectionQuality
   ) => Promise<AVPlaybackStatus>;
-  setVolumeAsync!: (volume: number) => Promise<AVPlaybackStatus>;
+  setVolumeAsync!: (volume: number, audioPan?: number) => Promise<AVPlaybackStatus>;
   setIsMutedAsync!: (isMuted: boolean) => Promise<AVPlaybackStatus>;
   setIsLoopingAsync!: (isLooping: boolean) => Promise<AVPlaybackStatus>;
   setProgressUpdateIntervalAsync!: (